# api/v2/swagger/swagger.yaml
definitions:
  api.BusinessError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  api.Response:
    properties:
      data:
        type: object
      response:
        $ref: '#/definitions/api.BusinessError'
    type: object
  combine.translationWithPatternPostReq:
    properties:
      combine:
        type: integer
      components:
        items:
          type: string
        type: array
      language:
        type: string
      productName:
        type: string
      region:
        type: string
      scope:
        type: string
      scopeFilter:
        type: string
      version:
        type: string
    required:
    - combine
    - components
    - language
    - productName
    - scope
    - version
    type: object
...
info:
  contact:
    name: Singleton Team
    url: https://github.com/vmware/singleton
  description: This is a Singleton server
  license:
    name: EPL-2.0
    url: https://www.eclipse.org/legal/epl-2.0/
  title: Singleton Service API
  version: "2.0"
paths:
  /combination/translationsAndPattern:
    get:
      description: Get translation and pattern data by customized type
      parameters:
      - description: an integer which represents combine type number 1 or 2
        in: query
        name: combine
        required: true
        type: integer
      - description: product name
        in: query
        name: productName
        required: true
        type: string
      - description: version
        in: query
        name: version
        required: true
        type: string
      - description: a string contains multiple components, separated by commas. e.g.
          'cim,common,cpa,cpu'
        in: query
        name: components
        required: true
        type: string
      - description: a string which represents language, e.g. en,en-US,pt,pt-BR,zh-Hans
        in: query
        name: language
        required: true
        type: string
      - description: a string which represents region, e.g. US,PT,CN
        in: query
        name: region
        type: string
      - description: pattern category string, separated by commas. e.g. 'dates,numbers,currencies,plurals,measurements,dateFields'
        in: query
        name: scope
        required: true
        type: string
      - description: a string for filtering the pattern data, separated by comma and
          underline. e.g. 'dates_eras,dates_dateTimeFormats'
        in: query
        name: scopeFilter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get translation and pattern data
      tags:
      - translation-with-pattern-api
    post:
      deprecated: true
      description: Get translation and pattern data by customized type
      parameters:
      - description: translationWithPatternPostReq
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/combine.translationWithPatternPostReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get translation and pattern data (Deprecated because GET method is
        ready)
      tags:
      - translation-with-pattern-api
  /formatting/date/localizedDate:
    get:
      deprecated: true
      description: Get localized date by locale and pattern
      parameters:
      - description: locale String. e.g. 'en-US'
        in: query
        name: locale
        required: true
        type: string
      - description: long value of the date(e.g. 1472728030290). the specified number
          of milliseconds since the standard base time known as 'the epoch', namely
          January 1, 1970, 00:00:00 GMT.
        in: query
        name: longDate
        required: true
        type: integer
      - description: 'pattern used to format the long date(the value could be one
          of this: YEAR = ''y'',QUARTER = ''QQQQ'',ABBR_QUARTER = ''QQQ'',QUARTER_YEAR
          = ''QQQQy'',QUARTER_ABBR_YEAR = ''QQQy'',MONTH = ''MMMM'',ABBR_MONTH = ''MMM'',NUM_MONTH
          = ''M'',MONTH_YEAR = ''MMMMy'',MONTH_ABBR_YEAR = ''MMMy'',MONTH_NUM_YEAR
          = ''My'',DAY = ''d'',MONTH_DAY_YEAR = ''MMMMdy'',ABBR_MONTH_DAY_YEAR = ''MMMdy'',NUM_MONTH_DAY_YEAR
          = ''Mdy'',WEEKDAY = ''EEEE'',ABBR_WEEKDAY = ''E'',WEEKDAY_MONTH_DAY_YEAR
          = ''EEEEMMMMdy'',ABBR_WEEKDAY_MONTH_DAY_YEAR = ''EMMMdy'',NUM_WEEKDAY_MONTH_DAY_YEAR
          = ''EMdy'',MONTH_DAY = ''MMMMd'',ABBR_MONTH_DAY = ''MMMd'',NUM_MONTH_DAY
          = ''Md'',WEEKDAY_MONTH_DAY = ''EEEEMMMMd'',ABBR_WEEKDAY_MONTH_DAY = ''EMMMd'',NUM_WEEKDAY_MONTH_DAY
          = ''EMd'')'
        in: query
        name: pattern
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get localized date
      tags:
      - formatting-api
...
  /locale/supportedLanguageList:
    get:
      description: Get language display names in a specified locale
      parameters:
      - description: product name
        in: query
        name: productName
        required: true
        type: string
      - description: version
        in: query
        name: version
        required: true
        type: string
      - description: displayLanguage
        in: query
        name: displayLanguage
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "206":
          description: Successful Partially
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get language display names
      tags:
      - locale-api
...
swagger: "2.0"
...